+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/connections.ts
@@ -8,9 +8,9 @@ import { mapConnectionsByDestination } from 'n8n-workflow';
 
 import type { SimpleWorkflow } from '@/types';
 
-import type { Violation } from '../types/evaluation';
-import type { SingleEvaluatorResult } from '../types/test-result';
-import { calcSingleEvaluatorScore } from '../utils/score';
+import type { Violation } from '../../types/evaluation';
+import type { SingleEvaluatorResult } from '../../types/test-result';
+import { calcSingleEvaluatorScore } from '../../utils/score';
 
 export function resolveConnections<T = INodeInputConfiguration>(
 	connections: Array<NodeConnectionType | T> | ExpressionString,
