+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/from-ai.ts
@@ -2,10 +2,10 @@ import type { INodeTypeDescription } from 'n8n-workflow';
 
 import type { SimpleWorkflow } from '@/types';
 
-import type { Violation } from '../types/evaluation';
-import type { SingleEvaluatorResult } from '../types/test-result';
-import { isTool } from '../utils/isTool';
-import { calcSingleEvaluatorScore } from '../utils/score';
+import type { Violation } from '../../types/evaluation';
+import type { SingleEvaluatorResult } from '../../types/test-result';
+import { isTool } from '../../utils/isTool';
+import { calcSingleEvaluatorScore } from '../../utils/score';
 
 /**
  * Checks if a value contains $fromAI or $fromAi reference
