+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/trigger.test.ts
